Доработка после проверки ДЗ
