+++ js/main.js
@@ -1,3 +1,6 @@
 import './data.js';
 import './util.js';
-import './thumbnail-rendering.js';
+import {renderThumbnails} from './thumbnail.js';
+import {makePosts} from './data.js';
+
+renderThumbnails(makePosts());
